requirements.txt
psycopg2-binary
pgvector
numpy